Module swiperGlobalFunctionality.js was created
