+++ js/script.js
@@ -4,3 +4,6 @@ handleBurgerButtons();
 import popupRegistration from "./modules/popupRegistration.js";
 popupRegistration();
 
+import { setSlidesState } from "./modules/swiperGlobalFunctionality.js";
+import { getActiveIndexes } from "./modules/swiperGlobalFunctionality.js";
+import { navigationButtonAppearance } from "./modules/swiperGlobalFunctionality.js";
